news-bot/requirements.txt
requests
praw
feedparser
python-dotenv
schedule
